检测相关代码，单目标评测优化

+++ src/com/lcf/util/common/BoxStruct.java
@@ -1,8 +1,8 @@
 package com.lcf.util.common;
 
 public class BoxStruct {
-	public int x;
-	public int y;
-	public int w;
-	public int h;
+	public double x;
+	public double y;
+	public double w;
+	public double h;
 }
